refactor : 테스트에서만 사용되는 클래스 패키지 이동
