+++ src/test/java/racingcar/testCondition/FalseCondition.java
@@ -1,4 +1,6 @@
-package racingcar.condition;
+package racingcar.testCondition;
+
+import racingcar.condition.MovingCarCondition;
 
 public class FalseCondition implements MovingCarCondition {
     @Override
