+++ src/test/java/racingcar/testCondition/TrueCondition.java
@@ -1,4 +1,6 @@
-package racingcar.condition;
+package racingcar.testCondition;
+
+import racingcar.condition.MovingCarCondition;
 
 public class TrueCondition implements MovingCarCondition {
     @Override
